+++ Libplanet/Net/Swarm.cs
@@ -44,6 +44,9 @@ namespace Libplanet.Net
         private CancellationTokenSource _workerCancellationTokenSource;
         private CancellationToken _cancellationToken;
 
+        private (long, BoundPeer, HashDigest<SHA256>)? _demandBlockHash;
+        private ConcurrentDictionary<TxId, BoundPeer> _demandTxIds;
+
         static Swarm()
         {
             if (!(Type.GetType("Mono.Runtime") is null))
@@ -269,6 +272,8 @@ namespace Libplanet.Net
             _cancellationToken = CancellationTokenSource.CreateLinkedTokenSource(
                     _workerCancellationTokenSource.Token, cancellationToken
                 ).Token;
+            _demandBlockHash = null;
+            _demandTxIds = new ConcurrentDictionary<TxId, BoundPeer>();
             await Transport.StartAsync(_cancellationToken);
 
             _logger.Debug("Starting swarm...");
@@ -278,6 +283,8 @@ namespace Libplanet.Net
             {
                 tasks.Add(Transport.RunAsync(_cancellationToken));
                 tasks.Add(BroadcastTxAsync(broadcastTxInterval, _cancellationToken));
+                tasks.Add(ProcessFillblock(_cancellationToken));
+                tasks.Add(ProcessFillTxs(_cancellationToken));
                 _logger.Debug("Swarm started.");
 
                 await await Task.WhenAny(tasks);
@@ -1141,17 +1148,13 @@ namespace Libplanet.Net
 
                 case TxIds txIds:
                     {
-                        Task.Run(
-                            async () => await ProcessTxIds(txIds, _cancellationToken),
-                            _cancellationToken);
+                        ProcessTxIds(txIds);
                         break;
                     }
 
                 case BlockHashes blockHashes:
                     {
-                        Task.Run(
-                            async () => await ProcessBlockHashes(blockHashes, _cancellationToken),
-                            _cancellationToken);
+                        _logger.Error("BlockHashes message is only for IBD.");
                         break;
                     }
 
@@ -1168,32 +1171,6 @@ namespace Libplanet.Net
             }
         }
 
-        private async Task ProcessBlockHashes(
-            BlockHashes message,
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            if (!(message.Remote is BoundPeer peer))
-            {
-                _logger.Information(
-                    "BlockHashes was sent from invalid peer " +
-                    $"[{message.Remote.Address.ToHex()}]. ignored.");
-                return;
-            }
-
-            ImmutableList<HashDigest<SHA256>> newHashes = message.Hashes
-                .Where(hash => !_store.ContainsBlock(hash))
-                .ToImmutableList();
-
-            if (newHashes.Any())
-            {
-                await FetchBlocks(peer, newHashes, cancellationToken);
-            }
-            else
-            {
-               _logger.Debug($"No blocks are required; {nameof(BlockHashes)} is ignored.");
-            }
-        }
-
         private async Task ProcessBlockHeader(
             BlockHeaderMessage message,
             CancellationToken cancellationToken = default(CancellationToken))
@@ -1201,80 +1178,40 @@ namespace Libplanet.Net
             if (!(message.Remote is BoundPeer peer))
             {
                 _logger.Information(
-                    "BlockHashes was sent from invalid peer " +
-                    $"[{message.Remote.Address.ToHex()}]. ignored.");
+                    "BlockHeaderMessage was sent from invalid peer " +
+                    "{PeerAddress}; ignored.",
+                    message.Remote.Address);
                 return;
             }
 
             BlockHeaderReceived.Set();
             BlockHeader header = message.Header;
 
-            // FIXME: this should be changed into total difficulty.
-            if (header.Index > BlockChain.Tip.Index)
-            {
-                await FetchBlocks(
-                    peer,
-                    new[] { new HashDigest<SHA256>(header.Hash.ToArray()) },
-                    cancellationToken);
-            }
-            else
-            {
-                _logger.Debug($"No blocks are required; {nameof(BlockHeaderMessage)} is ignored.");
-            }
-        }
-
-        private async Task FetchBlocks(
-            BoundPeer peer,
-            IReadOnlyCollection<HashDigest<SHA256>> hashes,
-            CancellationToken cancellationToken)
-        {
             _logger.Debug(
-                $"Trying to {nameof(GetBlocksAsync)}() using {{0}} hashes.",
-                hashes.Count);
+                "Block header of hash [{Hash}] (index: {Index}) received.",
+                ByteUtil.Hex(header.Hash),
+                header.Index,
+                Address.ToHex());
 
-            System.Collections.Async.IAsyncEnumerable<Block<T>> fetched = GetBlocksAsync(
-                peer,
-                hashes
-            );
-
-            List<Block<T>> blocks = await fetched.ToListAsync(
-                cancellationToken
-            );
-            _logger.Debug($"{nameof(GetBlocksAsync)}() complete.");
-
-            if (!blocks.Any())
+            using (await _blockSyncMutex.LockAsync(cancellationToken))
             {
-                _logger.Debug("No blocks fetched.");
-                return;
-            }
-
-            LastReceived = DateTimeOffset.UtcNow;
-            BlockReceived.Set();
-
-            try
-            {
-                using (await _blockSyncMutex.LockAsync(cancellationToken))
+                // FIXME: this should be changed into total difficulty.
+                if (header.Index > BlockChain.Tip.Index &&
+                    (_demandBlockHash is null || _demandBlockHash.Value.Item1 < header.Index))
                 {
-                    if (await AppendBlocksAsync(peer, blocks, cancellationToken))
-                    {
-                        BroadcastBlock(peer.Address, blocks.Last());
-                    }
-
-                    _logger.Debug("Append complete.");
+                    _demandBlockHash =
+                        (header.Index, peer, new HashDigest<SHA256>(header.Hash.ToArray()));
                 }
-            }
-            catch (TimeoutException)
-            {
-                // As we have more chances, ignore this.
-                _logger.Debug($"Timeout occurred during {nameof(ProcessBlockHashes)}().");
-            }
-            catch (Exception e)
-            {
-                _logger.Error(
-                    e,
-                    $"Append failed during {nameof(ProcessBlockHashes)}() due to exception: {{0}}",
-                    e);
-                throw;
+                else
+                {
+                    _logger.Debug(
+                        "No blocks are required " +
+                        "(current: {Current}, demand: {Demand}, received: {Received});" +
+                        $" {nameof(BlockHeaderMessage)} is ignored.",
+                        BlockChain.Tip.Index,
+                        _demandBlockHash?.Item1,
+                        header.Index);
+                }
             }
         }
 
@@ -1355,75 +1292,6 @@ namespace Libplanet.Net
                         blockChain.Swap(synced, evaluateActions);
                     }
                 }
-            }
-        }
-
-        private async Task<bool> AppendBlocksAsync(
-            BoundPeer peer,
-            List<Block<T>> blocks,
-            CancellationToken cancellationToken
-        )
-        {
-            // We assume that the blocks are sorted in order.
-            Block<T> oldest = blocks.First();
-            Block<T> latest = blocks.Last();
-            Block<T> tip = BlockChain.Tip;
-
-            if (tip is null || latest.Index > tip.Index)
-            {
-                List<Block<T>> blocksToAppend;
-                if (oldest.PreviousHash is null)
-                {
-                    _logger.Debug("The oldest block[{block}] seems to be genesis.", oldest);
-                    blocksToAppend = blocks;
-                }
-                else if (!(tip is null) &&
-                         blocks.Any(block => block.PreviousHash.Equals(tip.Hash)))
-                {
-                    // FIXME: This may not work as expected in multi-miner environment.
-                    _logger.Debug("Does not need to fill.");
-                    blocksToAppend = blocks.Where(block => block.Index > tip.Index).ToList();
-                }
-                else
-                {
-                    _logger.Debug("Trying to fill up previous blocks...");
-                    await SyncPreviousBlocksAsync(
-                        BlockChain,
-                        peer,
-                        oldest.PreviousHash,
-                        null,
-                        blocks.Count,
-                        evaluateActions: true,
-                        cancellationToken: cancellationToken
-                    );
-                    _logger.Debug("Filled up; trying to concatenation...");
-                    blocksToAppend = blocks;
-                }
-
-                foreach (Block<T> block in blocksToAppend)
-                {
-                    BlockChain.Append(block);
-                }
-
-                _logger.Debug("Sync is done.");
-
-                BlockAppended.Set();
-
-                var blockHashes =
-                    blocks.Aggregate(string.Empty, (current, block) =>
-                        current + $"[{block.Hash.ToString()}]");
-                _logger.Debug(
-                    $"Re-broadcast {nameof(BlockHashes)} with {{0}} blocks, which are {{1}}.",
-                    blocks.Count,
-                    blockHashes);
-                return true;
-            }
-            else
-            {
-                _logger.Information(
-                    "Received index is older than current chain's tip." +
-                    " ignored.");
-                return false;
             }
         }
 
@@ -1583,9 +1451,7 @@ namespace Libplanet.Net
             }
         }
 
-        private async Task ProcessTxIds(
-            TxIds message,
-            CancellationToken cancellationToken = default(CancellationToken))
+        private void ProcessTxIds(TxIds message)
         {
             if (!(message.Remote is BoundPeer peer))
             {
@@ -1595,12 +1461,12 @@ namespace Libplanet.Net
                 return;
             }
 
-            _logger.Debug("Trying to fetch txs...");
             _logger.Debug(
                 "Received TxIds: [{txIds}]",
                 string.Join(", ", message.Ids));
 
             ImmutableHashSet<TxId> newTxIds = message.Ids
+                .Where(id => !_demandTxIds.ContainsKey(id))
                 .Where(id => !_store.ContainsTransaction(id))
                 .ToImmutableHashSet();
 
@@ -1610,24 +1476,11 @@ namespace Libplanet.Net
                 return;
             }
 
-            List<Transaction<T>> txs;
-            try
+            _logger.Debug("Txs to require: {@txIds}", newTxIds.Select(txid => txid.ToString()));
+            foreach (var txid in newTxIds)
             {
-                System.Collections.Async.IAsyncEnumerable<Transaction<T>> fetched = GetTxsAsync(
-                    peer, newTxIds, cancellationToken);
-                txs = await fetched.ToListAsync(cancellationToken);
+                _demandTxIds.TryAdd(txid, peer);
             }
-            catch (TimeoutException)
-            {
-                _logger.Debug($"Timeout occurred during {nameof(ProcessTxIds)}().");
-                return;
-            }
-
-            BlockChain.StageTransactions(txs.ToImmutableHashSet());
-            TxReceived.Set();
-            _logger.Debug("Txs staged successfully.");
-
-            BroadcastTxs(message.Remote.Address, txs);
         }
 
         private void TransferBlocks(GetBlocks getData)
@@ -1817,6 +1670,117 @@ namespace Libplanet.Net
             };
 
             Transport.ReplyMessage(reply);
+        }
+
+        private async Task ProcessFillblock(CancellationToken cancellationToken)
+        {
+            while (!cancellationToken.IsCancellationRequested)
+            {
+                if (_demandBlockHash is null ||
+                    _demandBlockHash.Value.Item1 <= BlockChain.Tip.Index)
+                {
+                    await Task.Delay(100, cancellationToken);
+                    continue;
+                }
+
+                (long index, BoundPeer peer, HashDigest<SHA256> blockHash) =
+                    _demandBlockHash.Value;
+                try
+                {
+                    await SyncPreviousBlocksAsync(
+                        BlockChain,
+                        peer,
+                        blockHash,
+                        null,
+                        0,
+                        true,
+                        cancellationToken);
+
+                    // FIXME: Clean up events
+                    BlockReceived.Set();
+                    BlockAppended.Set();
+                    BroadcastBlock(peer.Address, BlockChain.Tip);
+                }
+                catch (TimeoutException)
+                {
+                    _logger.Debug($"Timeout occurred during {nameof(ProcessFillblock)}");
+                    await Task.Delay(100, cancellationToken);
+                }
+                catch (Exception e)
+                {
+                    var msg =
+                        $"Unexpected exception occurred during" +
+                        $" {nameof(ProcessFillblock)}: {{e}}";
+                    _logger.Error(e, msg, e);
+                }
+            }
+        }
+
+        private async Task ProcessFillTxs(CancellationToken cancellationToken)
+        {
+            while (!cancellationToken.IsCancellationRequested)
+            {
+                if (_demandTxIds.IsEmpty)
+                {
+                    await Task.Delay(100, cancellationToken);
+                    continue;
+                }
+
+                _logger.Debug(
+                    "Processing txids: {@txIds}",
+                    _demandTxIds.Keys.Select(txid => txid.ToString()));
+                var demandTxIds = _demandTxIds.ToArray();
+                var demands = new Dictionary<BoundPeer, List<TxId>>();
+
+                foreach (var kv in demandTxIds)
+                {
+                    if (!demands.ContainsKey(kv.Value))
+                    {
+                        demands[kv.Value] = new List<TxId>();
+                    }
+
+                    demands[kv.Value].Add(kv.Key);
+                }
+
+                var tasks = new List<Task<List<Transaction<T>>>>();
+                foreach (var kv in demands)
+                {
+                    System.Collections.Async.IAsyncEnumerable<Transaction<T>> fetched =
+                        GetTxsAsync(
+                            kv.Key, kv.Value, cancellationToken);
+                    tasks.Add(fetched.ToListAsync(cancellationToken));
+                }
+
+                var txs = new List<Transaction<T>>();
+                try
+                {
+                    await tasks.WhenAll();
+                }
+                catch (Exception)
+                {
+                    _logger.Information(
+                        $"Some tasks faulted during {nameof(GetTxsAsync)}().");
+                }
+
+                foreach (var task in tasks.Where(task => !task.IsFaulted))
+                {
+                    txs.AddRange(task.Result);
+                }
+
+                BlockChain.StageTransactions(txs.ToImmutableHashSet());
+                TxReceived.Set();
+                _logger.Debug(
+                    "Txs staged successfully: {@txIds}",
+                    txs.Select(tx => tx.Id.ToString()));
+
+                // FIXME: Should exclude peers of source of the transaction ids.
+                BroadcastTxs(null, txs);
+
+                foreach (var kv in demandTxIds)
+                {
+                    _demandTxIds.TryRemove(kv.Key, out BoundPeer value);
+                }
+            }
         }
     }
 }